Simplify collection selection, refactor some commands, add fma report
This change moves us away from a strict folder structure rooted at
$FMA_ROOT, with collections that need to belong to particular 'types'
and that can be listed as part of a whole.

Instead, commands can just operate on a particular collection, where a
collection is just a directory. A default collection can be saved via
the config for convenience, and a CollectionCommand base class can be
used for commands that operate on a particular collection.